+++ app/aides/exoneration-fiscale/denormandie/Denormandie.tsx
@@ -2,38 +2,12 @@
 import { Card } from '@/components/UI'
 import rules from '@/app/règles/rules'
 import checkIcon from '@/public/check.svg'
-import Publicodes, { formatValue } from 'publicodes'
-import useSetSearchParams from '@/components/useSetSearchParams'
-import { useSearchParams } from 'next/navigation'
-import {
-  encodeDottedName,
-  getAnsweredQuestions,
-  getSituation,
-} from '@/components/publicodes/situationUtils'
-import {
-  CommuneLogement,
-  DureeLocation,
-  Li,
-  MontantQuestion,
-  QuestionList,
-  TypeResidence,
-  YesNoQuestion,
-} from '@/app/module/AmpleurQuestions'
 import FatConseiller from '@/components/FatConseiller'
-import { EligibilityResult } from '@/components/EligibilityResult'
+import EligibilityDenormandie from '@/components/module/EligibilityDenormandie'
 
 export default function Denormandie() {
-  const engine = new Publicodes(rules)
   const dottedName = 'denormandie'
-  const setSearchParams = useSetSearchParams()
-  const rawSearchParams = useSearchParams(),
-    searchParams = Object.fromEntries(rawSearchParams.entries())
 
-  const situation = getSituation(searchParams, rules)
-  const answeredQuestions = getAnsweredQuestions(searchParams, rules)
-  const evaluation = engine
-    .setSituation(situation)
-    .evaluate(dottedName + ' . montant')
   return (
     <>
       <h2>Le dispositif Denormandie</h2>
@@ -50,126 +24,7 @@ export default function Denormandie() {
         />
       </div>
       <Card>
-        <h3
-          css={`
-            margin-top: 1rem;
-          `}
-        >
-          Etes-vous éligible au dispositif Denormandie ?
-        </h3>
-        <QuestionList>
-          <Li
-            $next={true}
-            $touched={answeredQuestions.includes(
-              'denormandie . commune . éligible',
-            )}
-          >
-            <CommuneLogement
-              {...{
-                setSearchParams,
-                situation,
-                answeredQuestions,
-                onChange: (result) => {
-                  setSearchParams({
-                    [encodeDottedName('logement . commune . denormandie')]:
-                      result.eligibilite['logement . commune . denormandie'] +
-                      '*',
-                    [encodeDottedName('logement . commune')]:
-                      `"${result.code}"*`,
-                    [encodeDottedName('logement . commune . nom')]:
-                      `"${result.nom}"*`,
-                  })
-                },
-              }}
-            />
-          </Li>
-          {situation['logement . commune . denormandie'] == 'oui' && (
-            <>
-              <Li
-                $next={true}
-                $touched={answeredQuestions.includes(
-                  'logement . résidence principale propriétaire',
-                )}
-              >
-                <TypeResidence
-                  {...{ setSearchParams, situation, answeredQuestions }}
-                />
-              </Li>
-              <Li
-                $next={answeredQuestions.includes(
-                  'logement . résidence principale propriétaire',
-                )}
-                $touched={answeredQuestions.includes("logement . prix d'achat")}
-              >
-                <MontantQuestion
-                  {...{
-                    setSearchParams,
-                    situation,
-                    answeredQuestions,
-                    rule: "logement . prix d'achat",
-                    text: "Le prix d'achat du logement est de",
-                  }}
-                />
-              </Li>
-              <Li
-                $next={answeredQuestions.includes("logement . prix d'achat")}
-                $touched={answeredQuestions.includes('projet . travaux')}
-              >
-                <MontantQuestion
-                  {...{
-                    setSearchParams,
-                    situation,
-                    answeredQuestions,
-                    rule: 'projet . travaux',
-                    text: 'Le budget travaux est estimé à',
-                  }}
-                />
-              </Li>
-              <Li
-                $next={answeredQuestions.includes('projet . travaux')}
-                $touched={answeredQuestions.includes(
-                  'denormandie . années de location',
-                )}
-              >
-                <DureeLocation
-                  {...{
-                    setSearchParams,
-                    situation,
-                    rules,
-                    answeredQuestions,
-                    rule: 'denormandie . années de location',
-                  }}
-                />
-              </Li>
-              <Li
-                $next={answeredQuestions.includes('projet . travaux')}
-                $touched={answeredQuestions.includes(
-                  'denormandie . amélioration énergétique',
-                )}
-              >
-                <YesNoQuestion
-                  {...{
-                    setSearchParams,
-                    situation,
-                    answeredQuestions,
-                    rules,
-                    rule: 'denormandie . amélioration énergétique',
-                    text: "Allez-vous améliorer la performance énergétique d'au moins 30% ou réaliser 2 types de travaux parmi: changement de chaudière, de production d'eau chaude, isolation thermique des combles, des murs ou des fenêtres.",
-                  }}
-                />
-              </Li>
-            </>
-          )}
-        </QuestionList>
-        <EligibilityResult
-          {...{
-            evaluation,
-            engine,
-            dottedName,
-            situation,
-            text: 'au dispositif Denormandie',
-          }}
-        />
+        <EligibilityDenormandie dottedName="denormandie" />
       </Card>
       <h3>Les principales conditions d'éligibilité ?</h3>
       <div
@@ -188,7 +43,7 @@ export default function Denormandie() {
       />
       <FatConseiller
         {...{
-          situation,
+          situation: {},
           margin: 'small',
           titre: 'Comment toucher cette aide ?',
           texte: rules[dottedName].commentFaireHtml,
