+++ app/module/denormandie/page.tsx
@@ -1,0 +1,20 @@
+'use client'
+import { Suspense } from 'react'
+import { FooterModule } from '../Ampleur'
+import { AmpleurWrapper } from '../AmpleurUI'
+import EligibilityDenormandie from '@/components/module/EligibilityDenormandie'
+
+export default function ModuleDenormandie() {
+  return (
+    <Suspense>
+      <AmpleurWrapper>
+        <EligibilityDenormandie
+          {...{
+            dottedName: 'denormandie',
+          }}
+        />
+        <FooterModule />
+      </AmpleurWrapper>
+    </Suspense>
+  )
+}
